chat room 수정
