+++ App.js
@@ -12,7 +12,7 @@ import Post_provide from './views/post/post_provide';
 import Post_ask from './views/post/post_ask';
 import SearchBar from './views/post/search_bar';
 import PostShow from './views/post/post_show';
-import ChatRoom from './views/chat/chat_room2';
+import ChatRoom from './views/chat/chat_room';
 import KakaoLogin from './views/login/kakao';
 import MyPgae_Location from './views/mypage/location';
 import Mypage_Like_List from './views/mypage/likeList';
